+++ src/components/DataActions.tsx
@@ -4,17 +4,17 @@ import { getFileHandle, saveToFile } from "@/utils/file";
 import { DataContext } from "@/context/data";
 import { modalContext } from "@/context/modal";
 import Tree from "./Tree";
-import Button, { behaviourEnum } from "./Button/Button";
+import Button from "./Button/Button";
 import ModalTemplateConfirm from "./Modal/ModalDefaults/ModalTemplateConfirm";
 import ModalPalette from "./Modal/ModalDefaults/ModalPalette";
-import LZString from "lz-string";
 import { toastContext } from "@/context/toast";
+import { ModalShare } from "./Modal/ModalDefaults/ModalShare";
 
 const DataActions: React.FC = () => {
   const { data, resetData, updateFileHandle, canEdit } =
     useContext(DataContext);
-  const { setContent } = useContext(modalContext);
-  const { showToastSuccess, showToastError } = useContext(toastContext);
+  const { setContent, closeModal } = useContext(modalContext);
+  const { showToastError, showToastSuccess } = useContext(toastContext);
 
   const openModalReset = () => {
     setContent(
@@ -24,14 +24,7 @@ const DataActions: React.FC = () => {
       >
         <span className="text paragraph">
           Resetting the notebook will erase all your content. If you want to
-          keep it, please{" "}
-          <Button
-            behaviour={behaviourEnum.POSITIVE}
-            onClick={() => saveToFile(data["RootPage"].title, data)}
-          >
-            Download
-          </Button>{" "}
-          it before resetting.
+          keep it, please download it before resetting.
         </span>
       </ModalTemplateConfirm>
     );
@@ -42,13 +35,23 @@ const DataActions: React.FC = () => {
   };
 
   const handleShare = () => {
-    const compressData = LZString.compressToEncodedURIComponent(
-      JSON.stringify(data)
+    setContent(<ModalShare />, false);
+  };
+
+  const handleGoHome = () => {
+    setContent(
+      <>
+        <p>Are you sure you want to leave this notebook?</p>
+        <Button
+          onClick={() => {
+            closeModal();
+            resetData();
+          }}
+        >
+          Confirm
+        </Button>
+      </>
     );
-    navigator.clipboard.writeText(
-      `${window.location.origin}?data=${compressData}`
-    );
-    showToastSuccess("Link copied to clipboard");
   };
 
   const handleLoad = async (e: any) => {
@@ -65,7 +68,7 @@ const DataActions: React.FC = () => {
     return (
       <div className={`${styles.dataActions} ${styles.grow_2}`}>
         <div className={styles.helper}>
-          <Button naked={true} onClick={resetData}>
+          <Button naked={true} onClick={handleGoHome}>
             <span className={styles["material-symbols-outlined"]}>home</span>
           </Button>
           <Button naked={true} onClick={handleUpdatePalette}>
@@ -93,7 +96,16 @@ const DataActions: React.FC = () => {
         </Button>
         <Button
           naked={true}
-          onClick={() => saveToFile(data["RootPage"].title, data)}
+          onClick={async () => {
+            const fileHandle = await saveToFile(data);
+            if (fileHandle) {
+              updateFileHandle(fileHandle);
+              showToastSuccess("Your data was successfuly saved. Now on, any change you make will be also saved localy!");
+            }
+            else {
+              showToastError("There was a problem saving your data")
+            }
+          }}
         >
           <span className={styles["material-symbols-outlined"]}>download</span>
         </Button>