+++ src/components/NoteList/NoteCard.tsx
@@ -30,13 +30,15 @@ const NoteCard: React.FC<IProps> = ({ itemKey, visible }) => {
         true
       )
     : "";
+
+  const [backgroundColor, color] = generateColor(key);
   return (
     <div
       id={`note-${key.split("_")[0]}`}
       className={`${styles.note} ${
         itemKey === selectedNote && styles.selected
       } ${visible && styles.visibleNote}`}
-      style={{ backgroundColor: generateColor(key) }}
+      style={{ backgroundColor, color }}
     >
       <div>
         <h2