+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## Version 1.0.0 (2024-XX-XX)
+
+### Features
+
+- Add temporal sharing through database
+- Changes are saved localy after download
+- Color text with high contrast colors
+
+### Fixes
+
 ## Version 0.1.5 (2024-07-21)
 
 ### Features